Start of a new Gathering controller
Want to reduce the number of actors, so we only really need an actor for talking with Twitter.  The counters can be relegated to a database which is accessed upon requests to the Gathering controller.  This controller just needs to look at the database (represented for now by the DataStore trait) and respond correctly.

  - Implemented start with a couple of tests